requirements.txt
streamlit==1.40.0
ultralytics==8.2.52
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84
streamlit-webrtc==0.49.0